+++ src/main/java/xyz/nikitacartes/easyauth/config/LangConfigV1.java
@@ -59,7 +59,8 @@ public class LangConfigV1 extends ConfigTemplate {
     public TranslatableText validSession = new TranslatableText("text.easyauth.validSession", "§aYou have a valid session. No need to log in.");
     public TranslatableText onlinePlayerLogin = new TranslatableText("text.easyauth.onlinePlayerLogin", "§aYou are using an online account. No need to log in.");
     public TranslatableText differentUsernameCase = new TranslatableText("text.easyauth.diffrentUsernameCase", "§6You are using a different case of your username. Please use the correct one.");
-
+    public TranslatableText wrongGlobalPassword = new TranslatableText("text.easyauth.wrongGlobalPassword", "§4Wrong global password!");
+    public TranslatableText registerRequiredWithGlobalPassword = new TranslatableText("text.easyauth.registerRequiredWithGlobalPassword", "§6Type /register <global password> <password> <password> to claim this account.");
 
     public LangConfigV1() {
         super("translation.conf");
@@ -113,6 +114,8 @@ public class LangConfigV1 extends ConfigTemplate {
         configValues.put("validSession", wrapIfNecessary(validSession));
         configValues.put("onlinePlayerLogin", wrapIfNecessary(onlinePlayerLogin));
         configValues.put("differentUsernameCase", wrapIfNecessary(differentUsernameCase));
+        configValues.put("wrongGlobalPassword", wrapIfNecessary(wrongGlobalPassword));
+        configValues.put("registerRequiredWithGlobalPassword", wrapIfNecessary(registerRequiredWithGlobalPassword));
 
         String configTemplate = Resources.toString(getResource("config/" + configPath), UTF_8);
         return new StringSubstitutor(configValues).replace(configTemplate);