+++ src/main/java/xyz/nikitacartes/easyauth/config/MainConfigV1.java
@@ -20,6 +20,7 @@ public class MainConfigV1 extends ConfigTemplate {
     public long resetLoginAttemptsTimeout = 120;
     public long sessionTimeout = 86400;
     public boolean enableGlobalPassword = false;
+    public boolean singleUseGlobalPassword = false;
     public boolean hidePlayerCoords = false;
     public boolean debug = false;
     public long configVersion = 2;
@@ -48,6 +49,7 @@ public class MainConfigV1 extends ConfigTemplate {
         configValues.put("resetLoginAttemptsTimeout", wrapIfNecessary(resetLoginAttemptsTimeout));
         configValues.put("sessionTimeout", wrapIfNecessary(sessionTimeout));
         configValues.put("enableGlobalPassword", wrapIfNecessary(enableGlobalPassword));
+        configValues.put("singleUseGlobalPassword", wrapIfNecessary(singleUseGlobalPassword));
         configValues.put("hidePlayerCoords", wrapIfNecessary(hidePlayerCoords));
         configValues.put("worldSpawn.dimension", wrapIfNecessary(worldSpawn.dimension));
         configValues.put("worldSpawn.x", wrapIfNecessary(worldSpawn.x));